cmake: update minimum cmake to 2.8
CMake >= 2.8.7 required for proper installation with GNUInstallDirs.
GNUInstallDirs really shouldn't effect earlier versions of CMake,
as OS's using old versions of cmake followed an older FHS. If
abosolutely necessary to be compatible with old CMake, I don't see
why we couldn't just bundle the latest GNUInstallDirs.cmake from
upstream with this application and only have it load in the case
of outdated CMake.