Correct the library linked in static build.
In static build, we should find the static pcre2 not the dynamic one.
For ease of maintenance, we also disable the shared library build when
static build is enabled.

Previously we escaped this error because CMake policy <= 3.2 does not
use full path to found linked library and thus the static pcre2 library
sneaks in.