Replace custom MSVC_MD option with standard CMAKE_MSVC_RUNTIME_LIBRARY. (#123)
* Replace custom MSVC_MD option with standard CMAKE_MSVC_RUNTIME_LIBRARY.

Instead of replacing /MD with /MT via search-and-replace, we now make
use of the standard CMAKE_MSVC_RUNTIME_LIBRARY option when building with
the build.ps1 script.

The CMAKE_MSVC_RUNTIME_LIBRARY option is supported since CMake 3.15 and
the root CMakeLists.txt file states CMake 3.16.3 as the minimum version,
so using this option should not be a problem.

* Pass CMAKE_MSVC_RUNTIME_LIBRARY to pcre2 as well